// planning-idsdp-pps-frontend/src/app/main/modules/rdpp-rtapp-management/features/dpp-tapp-list/project-details/models/project-details-partb.model.ts
export class ProjectDetailsPartB {
    uuid: string;
    projectConceptUuid: string;
    projectConceptMasterId: number;
    backgroundStatement: string;
    linkages: string;
    povertySituation: string;
    objective: string;
    outcomes: string;
    output: string;
    activities: string;
    sexDisaggregated: string;
    populationCoverage: string;
}